feat(workspace): Release packages in a workspace
This is effectively calling `cargo-release` for each package.  For
example, it does nothing to consolidate verson bump commits before or
after the release.

This uses `clap-cargo` to try to mirror `cargo`s logic for selecting
packages within a workspace.

This does have a bug wherein the packages are not released in dependency
order.

Fixes #66