feat(git): Allow a single commit to be used across a workspace
This adds a new configuration option, `consolidate-commits`.  When set in
the workspace, changes will be batched up until either all crates are
updated.

This also adds the concept of a "workspace-wide" configuration for
opterations like this.  In addition to `consolidate-commits`, the commit
message is read from the `workspace-wide` configuration.

As of right now, only `date` is supported as a commit message template
variable.  We could add the workspace root's crate name.  We probably
don't want the version from it because it might not match all of the
other crates.

Fixes #138