fix(workspace): Wait for publish to finish
A while back, publishing a crate was made asynchronous from
`cargo-publish`.  The problem is if we publish a crate at the bottom of
the dependency tree and then publish something that depends on it, that
can fail the test-compile because the crate is unavailable.  We didn't
run into problems before because usually we did enough stuff between
publish calls that it was done by the time we hit the next crate.
Switching to batch-processing broke this so now we are explicitly
waiting until the crate version is listed as published.

Fixes #182