fix(change): Improve accuracy of change-detection
Before, crates embedded in crates would flag the root crate as
changed. This now filters out those paths.

This could mislead people if they reach into other crates for files but
the risk seems low for that so I think this is safe enough.

Tested this with liquid which has a root crate and changed leaf crates.
The list of changed files for the root no longer lists other crates'
files in it.