fix(git): Read push-options for each crate
With workspaces, it was annoying to have a push per crate, so I merged
them together.  Then in #267, when implementing push options, I needed
something to read them from, so I made push config exclusively a
workspace-level features.

A better model for this is with `consolidate-commits` where people
opt-in to consolidation, and otherwise we use the crate's config.  This
will also help with #263

Fixes #269