fix(cargo): Don't panic on unexpected Cargo.tomls
When I first wrote this, `toml_edit`s API was more limited.  This
modernizes it to how I would write it now.  I assumed this would make it
shorter but apparently not, but I feel this clarifies intent more than
the repeated `contains_key` calls.

In refactoring this, the fix for #274 naturally fell out of it.

Fixes #274