fix: Don't treat bump as an exclude
Running
`cargo release --dry-run --no-dev-version --skip-publish --workspace --exclude utilities patch`

Treated `patch` as an `exclude` because of a naive use of the `clap`
crate.  This updates to the version of `clap_cargo` with the fix as well
as a whole slew of dependencies.

`clap_cargo`, `cargo_metadata`, `semver`, and `semver_parser` are in
lock-step with each other.  Thankfully, `semver` 1.0 is more like
the old `semver_parser` (before the merge `semver_parser` gained some
`semver` policy) so we could actually upgrade.  This did require
re-implementing some old `semver` logic because it was dropped.

Fixes #284
Fixes #249
Fixes #248