fix: Report excluded changes on 'downgrade'
In addressing #370, we left ourselves a hole where we might not log
changes (on absolute version which would be a downgrade for an excluded
package).

Ideally, we separate all these two code paths out so we can more
accurately detect all of this.  At this time, I'm not seeing a clean way
to do so, so instead we are loading data for excluded packages with no
version change, avoiding the error and allowing us to continue on.