fix: Correct publish order for circular dev dependencies
This is a fix for the following case. You have a workspace with one root
package and another sub-package. That sub-package could e.g. be a derive
macro implementation. In order to test that derive macro, you need some
functionality from the root packge. Therefore the derive macro has a
dev-dependency on the root package.

When publishing such a workspace, currently the way the order is
determined, is by looking at the dependencies between the workspace
packages. In this case the root depends on the sub-package and the other
way round. As the packages are looked at an arbitrary order, it is
determined that the root package should be published first and then the
sub-package. This isn't the correct order though.

This commit fixes that problem by excluding dev and build dependencies,
when looking at the dependencies within the workspace.