fix(cli): Properly log replacements
The replacement was not properly logged, making it nearly impossible
to figure out in which file a replacement failed, even with trace
logging.

To help with that, add a debug message with the full path of the file
where replacements are going to be tried in and in case replacement
fails, log the file path given in the config.