fix(commit): Skip if nothing changed
Especially with the new design, versions might already be applied,
replacements run, and hooks didn't change anything, so commit will fail
with an empty commit error.

We can't quite detect the above state, so rather than trying, I have the
commit call automatically skip the commit.  It still does it in dry-run
as the tree might still be clean.

Fixes #626